Adds coverage reports
Coverage report is generated by JaCoCo and sent to Coveralls via the
Maven plugin.

To allow compatibility with Surefire even when not running the JaCoCo
plugin, 'argLine' is set to empty on the parent POM and late evaluated
on modules that redefines the 'argLine' property (cogroo-ann and
cogroo-nlp).

An additional step in Travis CI configuration ensures that reports are
generated for every build.